Move render code to a module
