+++ include/render.h
@@ -1,0 +1,8 @@
+#ifndef RENDER_H
+# define RENDER_H
+
+# include "cub3d.h"
+
+void	render_view(t_cub3d *cub3d);
+
+#endif
